+++ snuba_sdk/__init__.py
@@ -2,12 +2,12 @@ from snuba_sdk.aliased_expression import AliasedExpression
 from snuba_sdk.column import Column
 from snuba_sdk.conditions import And, BooleanCondition, BooleanOp, Condition, Op, Or
 from snuba_sdk.entity import Entity
-from snuba_sdk.expressions import Granularity, Limit, Offset
-from snuba_sdk.flags import Consistent, Debug, Feature, ParentAPI, Team, Totals, Turbo
+from snuba_sdk.expressions import Granularity, Limit, Offset, Totals
 from snuba_sdk.function import CurriedFunction, Function, Identifier, Lambda
 from snuba_sdk.orderby import Direction, LimitBy, OrderBy
 from snuba_sdk.query import Query
 from snuba_sdk.relationships import Join, Relationship
+from snuba_sdk.request import Flags, Request
 
 __all__ = [
     "AliasedExpression",
@@ -16,12 +16,10 @@ __all__ = [
     "BooleanOp",
     "Column",
     "Condition",
-    "Consistent",
     "CurriedFunction",
-    "Debug",
     "Direction",
     "Entity",
-    "Feature",
+    "Flags",
     "Function",
     "Granularity",
     "Identifier",
@@ -33,10 +31,8 @@ __all__ = [
     "Op",
     "Or",
     "OrderBy",
-    "ParentAPI",
     "Query",
     "Relationship",
-    "Team",
+    "Request",
     "Totals",
-    "Turbo",
 ]